unit tests for us27 & us28
